Add validation during private key import
